Process discovered reader
